+++ packages/plugin-router/src/types/index.ts
@@ -3,35 +3,7 @@ import {
   RouteComponentProps,
 } from 'react-router-dom';
 import { History } from 'history';
-
-interface IModifyRoutes {
-  (modifyFn: IModifyFn): void;
-}
-
-interface IRouteWrapper {
-  (props: any): React.ComponentType<any>;
-}
-
-interface IDynamicImportComponent {
-  __LAZY__: boolean;
-  dynamicImport: () => Promise<{ default: React.ComponentType<any> }>;
-}
-
-interface RouteItemProps extends DefaultRouteProps {
-  children?: RouteItemProps[];
-  // disable string[]
-  path?: string;
-  // for rediect ability
-  redirect?: string;
-
-  component?: React.ComponentType<RouteComponentProps<any>> | React.ComponentType<any> | IDynamicImportComponent;
-
-  routeWrappers?: IRouteWrapper[];
-};
-
-interface IModifyFn {
-  (routes: RouteItemProps[]): RouteItemProps[];
-}
+import { RouteItemProps, IModifyRoutes } from './base';
 
 export interface IAppRouterProps {
   type?: 'hash' | 'browser' | 'memory';
