+++ packages/plugin-router/src/types/router.ts
@@ -1,30 +1,6 @@
-import {
-  RouteProps as DefaultRouteProps,
-  RouteComponentProps,
-} from 'react-router-dom';
 import { IgnoreOptions } from './collector';
 import { IRouterConfig } from '.';
-
-export interface IDynamicImportComponent {
-  __LAZY__: boolean;
-  dynamicImport: () => Promise<{ default: React.ComponentType<any> }>;
-}
-
-export interface IRouteWrapper {
-  (props: any): React.ComponentType<any>;
-}
-
-export interface RouteItemProps extends DefaultRouteProps {
-  children?: RouteItemProps[];
-  // disable string[]
-  path?: string;
-  // for rediect ability
-  redirect?: string;
-
-  component?: React.ComponentType<RouteComponentProps<any>> | React.ComponentType<any> | IDynamicImportComponent;
-
-  routeWrappers?: IRouteWrapper[];
-};
+import { RouteItemProps } from './base';
 
 export interface RouterProps {
   // custom props
@@ -43,10 +19,6 @@ export interface RouterProps {
   initialIndex?: number;
   fallback?: React.ReactNode;
 };
-
-export interface IModifyFn {
-  (routes: RouteItemProps[]): RouteItemProps[];
-}
 
 export interface RoutesProps {
   routes: IRouterConfig[];
